--- openhands/server/conversation_utils/__init__.py
@@ -1,1 +1,0 @@
-"""Server utilities."""
